pyproject.toml
[project]
name = "super-version"
version = "0.1.0"
description = ""
authors = [
    {name = "Achraf CHADIL",email = "chadil.hef@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[openai,silero,turn-detector] (==1.2.8)",
    "livekit-plugins-noise-cancellation (>=0.2,<1.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "aiocache (>=0.12.3,<0.13.0)",
    "chromadb (>=1.1.0,<2.0.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "uvloop (>=0.21.0,<0.22.0)",
    "httptools (>=0.6.4,<0.7.0)",
    "bleach (>=6.2.0,<7.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
ruff = "^0.13.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 150
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
lint.ignore = []
lint.fixable = ["ALL"]
lint.unfixable = []



[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
